requirements.txt
admin==0.0.1
appdirs==1.4.4
asgiref==3.2.10
attrs==20.2.0
autobahn==20.7.1
Automat==20.2.0
beautifulsoup4==4.9.1
bleach==3.2.1
certifi==2020.6.20
cffi==1.14.3
channels==2.4.0
chardet==3.0.4
colour==0.1.5
constantly==15.1.0
cryptography==3.1.1
cssselect2==0.3.0
daphne==2.5.0
Django==3.1.2
django-bleach==0.6.1
django-ckeditor==5.9.0
django-colorfield==0.3.2
django-common-utils==1.7.2
django-cors-headers==3.6.0
django-crontab==0.7.1
django-fernet-fields==0.6
django-filter==2.3.0
django-grip==3.0.0
django-hint==0.1.5
django-js-asset==1.2.2
django-lifecycle==0.7.7
django-object-actions==3.0.1
django-private-storage==2.2.2
django-rest-passwordreset==1.1.0
django-simple-email-confirmation==0.70
django-simple-history==2.11.0
django-user-sessions==1.7.1
djangorestframework==3.11.1
drf-nested-routers==0.92
drf-writable-nested==0.6.1
et-xmlfile==1.0.1
fcm-django==0.3.7
future==0.18.2
gender-guesser==0.4.0
gripcontrol==4.0.0
htmlmin==0.1.12
httpagentparser==1.9.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
iniconfig==1.1.1
jdcal==1.4.1
lorem==0.1.1
lxml==4.5.2
Markdown==3.2.2
mock==4.0.2
names==0.3.0
numpy
openpyxl==3.0.5
packaging==20.4
pdfminer.six==20200726
Pillow==7.2.0
pluggy==0.13.1
psycopg2-binary==2.8.6
pubcontrol==3.0.0
pwned-passwords-django==1.4.1
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydicti==1.1.3
pyfcm==1.4.9
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
PySocks==1.7.1
pytest==6.1.1
pytest-asyncio==0.14.0
pytest-django==3.10.0
python-dateutil==2.8.1
python-docx==0.8.10
python-dotenv==0.14.0
python-magic==0.4.18
python-pptx==0.6.18
pytz==2020.1
reportlab==3.5.53
requests==2.24.0
requests-toolbelt==0.9.1
secure-file-detection==0.1.5
service-identity==18.1.0
six==1.15.0
sortedcontainers==2.2.2
soupsieve==2.0.1
sqlparse==0.3.1
sseclient==0.0.27
stem==1.8.0
svglib==1.0.1
tinycss2==1.0.2
toml==0.10.1
torrequest==0.1.0
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.10
urlman==1.3.0
vidhdr==1.0.1
webencodings==0.5.1
Werkzeug==0.16.1
XlsxWriter==1.3.7
zope.interface==5.1.2
